+++ internal/annotations/pod.go
@@ -14,4 +14,9 @@ const (
 	// pod deletion order.
 	// The implicit deletion cost for pods that don't set the annotation is 0, negative values are permitted.
 	PodDeletionCost = "slinky.slurm.net/pod-deletion-cost"
+
+	// PodDeadline stores a time.RFC3339 timestamp, indicating when the Slurm node should complete its running
+	// workload by. Pods an earlier daedline are preferred to be deleted before pods with a later deadline.
+	// NOTE: this is honored on a best-effort basis, and does not offer guarantees on pod deletion order.
+	PodDeadline = "slinky.slurm.net/pod-deadline"
 )