+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Added container image labels.
 - Added `NodeSet.PersistentVolumeClaimRetentionPolicy.WhenScaled`
 - Added out-of-order scale-in for NodeSet pods.
+- Added NodeSet pod scale-in to consider running Slurm jobs.
 
 ### Fixed